# 7264: Improve the performance of the JFR parser