When parsing a JFR file, StackFrames are the objects that are the most
deserialized using Readers.
By default, JMC core parser used a generic ReflectiveReaders
which use reflection to parser the structure. However, we have known
objects like stack frames where we could directly map to the typed
object instead on relying on reflection, which speed up significantly
the parsing.

Here the results for c5.2xlarge instance single threaded parsing with 24MB jfr recording with 100 iterations:

| runs | baseline | specific readers optim | improvement |
| --- | --- | --- | --- |
| run 1 | 54,317ms | 39,623ms | -27% |
| run 2 | 54,988ms | 39,520ms | -28% |
| run 3 | 54,209ms | 39,904ms | -26% |

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JMC-7264](https://bugs.openjdk.java.net/browse/JMC-7264): Improve the performance of the JFR parser


### Reviewers
 * [Erik Gahlin](https://openjdk.java.net/census#egahlin) (@egahlin - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jmc pull/266/head:pull/266` \
`$ git checkout pull/266`

Update a local copy of the PR: \
`$ git checkout pull/266` \
`$ git pull https://git.openjdk.java.net/jmc pull/266/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 266`

View PR using the GUI difftool: \
`$ git pr show -t 266`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jmc/pull/266.diff">https://git.openjdk.java.net/jmc/pull/266.diff</a>

</details>