@egahlin 2021-06-14
I never thought reflection was a good idea, and It makes sense to have specialised logic for known types, but I don't think the code should rely on the field order.

In earlier release of JMC (4.1), there was similar specialiced logic as yours, but it broke once the "truncated" field was added to the stack trace struct in JDK 7. The same thing can happen again. 

Back then, there was a fields accessor class where the index was looked up when the parser was created. When the value was read the accessor could use the pre-calculated index to get the correct value. It might even have been bytecode generated for optimal performance.