@jpbempel 2021-06-14
Agreed that this code is fragile.
I can precompute the index and make sure readers are meant for the field at construction time.
I will try that.